fix: make exit status an enum, start moving testable code out of index
